仁慈阶段OK； 优化GB进场出场曲线动画以及淡入； 为圆形骨刺增加pitch扰动随着因子增加而变大； 仁慈攻击圆形骨刺扰动设置为50；
问题：
SansAi，二阶段时，有时候连击会丢失动态工厂，导致无法执行动态工厂的攻击；

待做：特殊阶段

+++ src/main/java/com/fanxing/fx_undertale/entity/capability/Animatable.java
@@ -9,9 +9,9 @@ import net.neoforged.neoforge.network.PacketDistributor;
  * @since 2025-11-20 20:58
  */
 public interface Animatable {
-    byte getAnimID();
-    void setAnimID(byte id);
-    default void sendAnimPacket(byte id){
+    int getAnimID();
+    void setAnimID(int id);
+    default void sendAnimPacket(int id){
         Entity entity = (Entity) this;
         PacketDistributor.sendToPlayersTrackingEntity(entity,new AnimPacket(entity.getId(),id,1f));
     }
